Moved Game from Home to Game page for now
