+++ src/App.tsx
@@ -1,6 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { NavBar } from "./components/NavBar";
 import { HomePage } from "./pages/HomePage";
+import GamePage from "./pages/GamePage";
 import { AppThemeProvider } from "./providers/ThemeProvider";
 
 export const App = () => {
@@ -10,6 +11,7 @@ export const App = () => {
         <NavBar />
         <Routes>
           <Route path="/" element={<HomePage />} />
+          <Route path="/game" element={<GamePage />} />
         </Routes>
       </BrowserRouter>
     </AppThemeProvider>
