+++ src/components/NavBar.tsx
@@ -35,6 +35,9 @@ export const NavBar: React.FC = () => {
         <NavItem>
           <AnimatedLink to="/speakers">Speakers</AnimatedLink>
         </NavItem>
+        <NavItem>
+          <AnimatedLink to="/game">Game</AnimatedLink>
+        </NavItem>
       </NavList>
       <Sidebar isOpen={isSidebarOpen}>
         <CloseIcon onClick={toggleSidebar}>&times;</CloseIcon>
@@ -56,6 +59,11 @@ export const NavBar: React.FC = () => {
           </SidebarItem>
           <SidebarItem>
             <SidebarLink to="/speakers" onClick={toggleSidebar}>
+              Speakers
+            </SidebarLink>
+          </SidebarItem>
+          <SidebarItem>
+            <SidebarLink to="/game" onClick={toggleSidebar}>
               Speakers
             </SidebarLink>
           </SidebarItem>
